+++ packages/notices/src/store/test/actions.js
@@ -17,15 +17,10 @@ describe( 'actions', () => {
 		const status = 'status';
 		const content = 'my message';
 
-		it( 'yields actions when options is empty', () => {
+		it( 'returns an action when options is empty', () => {
 			const result = createNotice( status, content );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: content,
-			} );
-
-			expect( result.next().value ).toMatchObject( {
+			expect( result ).toMatchObject( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {
@@ -41,12 +36,7 @@ describe( 'actions', () => {
 		it( 'normalizes content to string', () => {
 			const result = createNotice( status, <strong>Hello</strong> );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: expect.any( String ),
-			} );
-
-			expect( result.next().value ).toMatchObject( {
+			expect( result ).toMatchObject( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {
@@ -59,7 +49,7 @@ describe( 'actions', () => {
 			} );
 		} );
 
-		it( 'yields actions when options passed', () => {
+		it( 'returns an action when options passed', () => {
 			const context = 'foo';
 			const options = {
 				id,
@@ -69,25 +59,21 @@ describe( 'actions', () => {
 
 			const result = createNotice( status, content, options );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: content,
-			} );
-
-			expect( result.next().value ).toEqual( {
+			expect( result ).toEqual( {
 				type: 'CREATE_NOTICE',
 				context,
 				notice: {
 					id,
 					status,
 					content,
+					spokenMessage: content,
 					isDismissible: false,
 					actions: [],
 				},
 			} );
 		} );
 
-		it( 'yields action when speak disabled', () => {
+		it( 'returns an action when speak disabled', () => {
 			const result = createNotice(
 				undefined,
 				'my <strong>message</strong>',
@@ -99,13 +85,14 @@ describe( 'actions', () => {
 				}
 			);
 
-			expect( result.next().value ).toEqual( {
+			expect( result ).toEqual( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {
 					id,
 					status: DEFAULT_STATUS,
 					content: 'my <strong>message</strong>',
+					spokenMessage: null,
 					__unstableHTML: true,
 					isDismissible: false,
 					actions: [],
@@ -120,12 +107,7 @@ describe( 'actions', () => {
 
 			const result = createSuccessNotice( content );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: content,
-			} );
-
-			expect( result.next().value ).toMatchObject( {
+			expect( result ).toMatchObject( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {
@@ -144,12 +126,7 @@ describe( 'actions', () => {
 
 			const result = createInfoNotice( content );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: content,
-			} );
-
-			expect( result.next().value ).toMatchObject( {
+			expect( result ).toMatchObject( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {
@@ -168,12 +145,7 @@ describe( 'actions', () => {
 
 			const result = createErrorNotice( content );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: content,
-			} );
-
-			expect( result.next().value ).toMatchObject( {
+			expect( result ).toMatchObject( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {
@@ -192,12 +164,7 @@ describe( 'actions', () => {
 
 			const result = createWarningNotice( content );
 
-			expect( result.next().value ).toMatchObject( {
-				type: 'SPEAK',
-				message: content,
-			} );
-
-			expect( result.next().value ).toMatchObject( {
+			expect( result ).toMatchObject( {
 				type: 'CREATE_NOTICE',
 				context: DEFAULT_CONTEXT,
 				notice: {