+++ packages/notices/src/store/test/reducer.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import deepFreeze from 'deep-freeze';
-import { matchesProperty, find } from 'lodash';
 
 /**
  * Internal dependencies
@@ -12,11 +11,6 @@ import { createNotice, removeNotice } from '../actions';
 import { getNotices } from '../selectors';
 import { DEFAULT_CONTEXT } from '../constants';
 
-const getYieldedOfType = ( generatorAction, type ) => find(
-	Array.from( generatorAction ),
-	matchesProperty( [ 'type' ], type )
-);
-
 describe( 'reducer', () => {
 	it( 'should default to an empty object', () => {
 		const state = reducer( undefined, {} );
@@ -25,7 +19,7 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should track a notice', () => {
-		const action = getYieldedOfType( createNotice( 'error', 'save error' ), 'CREATE_NOTICE' );
+		const action = createNotice( 'error', 'save error' );
 		const state = reducer( undefined, action );
 
 		expect( state ).toEqual( {
@@ -33,6 +27,7 @@ describe( 'reducer', () => {
 				{
 					id: expect.any( String ),
 					content: 'save error',
+					spokenMessage: 'save error',
 					status: 'error',
 					isDismissible: true,
 					actions: [],
@@ -42,7 +37,7 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should track a notice by context', () => {
-		const action = getYieldedOfType( createNotice( 'error', 'save error', { context: 'foo' } ), 'CREATE_NOTICE' );
+		const action = createNotice( 'error', 'save error', { context: 'foo' } );
 		const state = reducer( undefined, action );
 
 		expect( state ).toEqual( {
@@ -50,6 +45,7 @@ describe( 'reducer', () => {
 				{
 					id: expect.any( String ),
 					content: 'save error',
+					spokenMessage: 'save error',
 					status: 'error',
 					isDismissible: true,
 					actions: [],
@@ -59,10 +55,10 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should track notices, respecting order by which they were created', () => {
-		let action = getYieldedOfType( createNotice( 'error', 'save error' ), 'CREATE_NOTICE' );
+		let action = createNotice( 'error', 'save error' );
 		const original = deepFreeze( reducer( undefined, action ) );
 
-		action = getYieldedOfType( createNotice( 'success', 'successfully saved' ), 'CREATE_NOTICE' );
+		action = createNotice( 'success', 'successfully saved' );
 		const state = reducer( original, action );
 
 		expect( state ).toEqual( {
@@ -70,6 +66,7 @@ describe( 'reducer', () => {
 				{
 					id: expect.any( String ),
 					content: 'save error',
+					spokenMessage: 'save error',
 					status: 'error',
 					isDismissible: true,
 					actions: [],
@@ -77,6 +74,7 @@ describe( 'reducer', () => {
 				{
 					id: expect.any( String ),
 					content: 'successfully saved',
+					spokenMessage: 'successfully saved',
 					status: 'success',
 					isDismissible: true,
 					actions: [],
@@ -86,7 +84,7 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should omit a removed notice', () => {
-		const action = getYieldedOfType( createNotice( 'error', 'save error' ), 'CREATE_NOTICE' );
+		const action = createNotice( 'error', 'save error' );
 		const original = deepFreeze( reducer( undefined, action ) );
 		const id = getNotices( original )[ 0 ].id;
 
@@ -98,7 +96,7 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should omit a removed notice by context', () => {
-		const action = getYieldedOfType( createNotice( 'error', 'save error', { context: 'foo' } ), 'CREATE_NOTICE' );
+		const action = createNotice( 'error', 'save error', { context: 'foo' } );
 		const original = deepFreeze( reducer( undefined, action ) );
 		const id = getNotices( original, 'foo' )[ 0 ].id;
 
@@ -110,7 +108,7 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should omit a removed notice across contexts', () => {
-		const action = getYieldedOfType( createNotice( 'error', 'save error' ), 'CREATE_NOTICE' );
+		const action = createNotice( 'error', 'save error' );
 		const original = deepFreeze( reducer( undefined, action ) );
 		const id = getNotices( original )[ 0 ].id;
 
@@ -120,10 +118,10 @@ describe( 'reducer', () => {
 	} );
 
 	it( 'should dedupe distinct ids, preferring new', () => {
-		let action = getYieldedOfType( createNotice( 'error', 'save error (1)', { id: 'error-message' } ), 'CREATE_NOTICE' );
+		let action = createNotice( 'error', 'save error (1)', { id: 'error-message' } );
 		const original = deepFreeze( reducer( undefined, action ) );
 
-		action = getYieldedOfType( createNotice( 'error', 'save error (2)', { id: 'error-message' } ), 'CREATE_NOTICE' );
+		action = createNotice( 'error', 'save error (2)', { id: 'error-message' } );
 		const state = reducer( original, action );
 
 		expect( state ).toEqual( {
@@ -131,6 +129,7 @@ describe( 'reducer', () => {
 				{
 					id: 'error-message',
 					content: 'save error (2)',
+					spokenMessage: 'save error (2)',
 					status: 'error',
 					isDismissible: true,
 					actions: [],