+++ packages/notices/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Master
+
+### Breaking Change
+
+- A notices message is no longer spoken as a result of notice creation, but rather by its display in the interface by its corresponding [`Notice` component](https://github.com/WordPress/gutenberg/tree/master/packages/components/src/notice).
+
 ## 1.1.2 (2019-01-03)
 
 ## 1.1.1 (2018-12-12)